# Conflicts:
#	src/main/java/Facultad/TrabajoPracticoDesarrollo/Controllers/HuespedController.java

+++ src/main/java/Facultad/TrabajoPracticoDesarrollo/DTOs/DtoTarjetaCredito.java
@@ -14,7 +14,7 @@ public class DtoTarjetaCredito extends DtoMedioPago {
     private String numeroDeTarjeta;
     private Date fechaVencimiento;
     private int codigoSeguridad;
-    private float monto;
+    private Double monto;
     private Moneda moneda;
     private Date fechaDePago;
     // Getters y Setters ... (omitidos, estándar)
@@ -41,7 +41,7 @@ public class DtoTarjetaCredito extends DtoMedioPago {
         private String numeroDeTarjeta;
         private Date fechaVencimiento;
         private int codigoSeguridad;
-        private float monto;
+        private Double monto;
         private Moneda moneda;
         private Date fechaDePago;
         private int cuotasCantidad;
@@ -54,7 +54,7 @@ public class DtoTarjetaCredito extends DtoMedioPago {
         public Builder numero(String val) { numeroDeTarjeta = val; return this; }
         public Builder vencimiento(Date val) { fechaVencimiento = val; return this; }
         public Builder seguridad(int val) { codigoSeguridad = val; return this; }
-        public Builder monto(float val) { monto = val; return this; }
+        public Builder monto(Double val) { monto = val; return this; }
         public Builder moneda(Moneda val) { moneda = val; return this; }
         public Builder fechaPago(Date val) { fechaDePago = val; return this; }
         public Builder cuotas(int val) { cuotasCantidad = val; return this; }
